move adaptive threshold module
