+++ nicegold_v5/__init__.py
@@ -61,6 +61,13 @@ from .qa import (
 from .rl_agent import RLScalper
 from .meta_classifier import MetaClassifier
 from .ml_dataset_m1 import generate_ml_dataset_m1
+from .adaptive_threshold_dl import (
+    ThresholdDataset,
+    ThresholdPredictor,
+    load_wfv_training_data,
+    train_threshold_predictor,
+    predict_thresholds,
+)
 
 try:
     from .deep_model_m1 import LSTMClassifier
