+++ nicegold_v5/adaptive_threshold_dl.py
